requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
brevo-python==1.1.2
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-multiselectfield==0.1.13
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.1
ecdsa==0.19.1
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
MarkupSafe==3.0.2
minio==7.2.15
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.23.1
sendgrid==6.12.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
